requirements.txt
streamlit==0.90.0
pdfminer.six==20201018
nltk==3.6.3
wordcloud==1.8.1
matplotlib==3.4.3
seaborn==0.11.2
pandas==1.3.3
